+++ src/components/Window.js
@@ -12,7 +12,7 @@ export default class Window extends PureComponent {
 
   window = null;
 
-  getFeaturesString = () =>
+  _getFeaturesString = () =>
     Object.entries({
       top: this.props.y,
       left: this.props.x,
@@ -27,38 +27,58 @@ export default class Window extends PureComponent {
       .map(([key, value]) => `${key}=${value}`)
       .join();
 
-  onBeforeUnload = () => {
+  _onBeforeUnload = () => {
     const { onClose } = this.props;
     if (onClose) {
       onClose(this.window);
     }
   };
 
+  _setupDom = () => {
+    const { stylesheet, title } = this.props;
+    if (typeof stylesheet === "string") {
+      if (stylesheet.search(/https?:\/\//) === -1) {
+        console.warn("If stylesheet is provided, must be an absolute URL");
+      } else {
+        var css = document.createElement("link");
+        css.setAttribute("rel", "stylesheet");
+        css.setAttribute("href", stylesheet);
+        this.window.document.head.appendChild(css);
+      }
+    } else {
+      const parent = this.window.opener;
+      parent.document.querySelectorAll("style").forEach(styleEl => {
+        this.window.document.head.appendChild(styleEl.cloneNode(true));
+      });
+    }
+    const rootEl = document.createElement("div");
+    this.window.document.title = title;
+    this.window.document.body.appendChild(rootEl);
+    this.window.addEventListener("beforeunload", this._onBeforeUnload, {
+      once: true,
+    });
+
+    return rootEl;
+  };
+
   componentDidMount() {
-    const { url, name, title, onBlock, onLoad } = this.props;
-    this.window = window.open(url, name, this.getFeaturesString());
+    const { url, name, onBlock, onLoad } = this.props;
+    this.window = window.open(url, name, this._getFeaturesString());
 
     if (!this.window) {
       if (onBlock) {
         onBlock();
       }
     } else {
-      this.window.document.title = title;
-      const rootEl = document.createElement("div");
-      this.window.document.body.appendChild(rootEl);
-      this.window.addEventListener("beforeunload", this.onBeforeUnload, {
-        once: true,
-      });
-
       // NOTE: Reason to use unstable_renderSubtreeIntoContainer:
       // https://twitter.com/dan_abramov/status/774591045980024833?lang=en
       // https://github.com/facebook/react/issues/4819
       unstable_renderSubtreeIntoContainer(
-        this, 
-        typeof this.props.children === 'function' 
-          ? this.props.children(this.window) 
-          : this.props.children, 
-        rootEl
+        this,
+        typeof this.props.children === "function"
+          ? this.props.children(this.window)
+          : this.props.children,
+        this._setupDom()
       );
 
       if (onLoad) {